Create server.js, move projectModel into /projects, move actionModel into /actions and create create routers for both projects and actions into corresponding directory
