+++ actions/actionsRouter.js
@@ -1,0 +1,5 @@
+const express = require('express');
+const Actions = require('./actionModel');
+const actionsRouter = express.Router();
+
+module.exports = actionsRouter;
