+++ projects/projectsRouter.js
@@ -1,0 +1,5 @@
+const express = require('express');
+const Projects = require('./projectModel');
+const projectsRouter = express.Router();
+
+module.exports = projectsRouter;
